fix(ci): maintainability gate — split test_workspace_retry.py under the 1500-line limit
The 1584-line tests/unit/service/test_workspace_retry.py tripped the
first-party file line-limit guardrail. Extract shared fixtures/helpers
into _workspace_retry_helpers.py and move the salvage/metadata/planning
retry tests into test_workspace_retry_salvage.py, keeping the preflight
tests in the original module.
